+++ browser/modules/anchor.js
@@ -12,6 +12,7 @@ let schema = urlparser.schema;
 let cloud, layers, setBaseLayer;
 let _self = false;
 let initMapParameters;
+let utils;
 
 /**
  *
@@ -22,6 +23,7 @@ module.exports = {
         cloud = o.cloud;
         layers = o.layers;
         setBaseLayer = o.setBaseLayer;
+        utils = o.utils;
         return this;
     },
     init: function () {
@@ -71,7 +73,13 @@ module.exports = {
         return result;
     },
 
-    applyMapParameters: (parameters) => {
+    applyMapParameters: (parameters, ignoreInitZoomCenter) => {
+        const arr = window.vidiConfig?.initZoomCenter ? utils.parseZoomCenter(window.vidiConfig.initZoomCenter) : null;
+        if (arr && !ignoreInitZoomCenter) {
+            parameters.zoom = arr.z;
+            parameters.x = arr.x;
+            parameters.y = arr.y;
+        }
         let result = new Promise((resolve, reject) => {
             if (parameters.x && parameters.y && parameters.zoom) {
                 cloud.get().setView(new L.LatLng(parseFloat(parameters.y), parseFloat(parameters.x)), parameters.zoom);
@@ -87,7 +95,6 @@ module.exports = {
                 reject();
             });
         });
-
         return result;
     },