+++ browser/modules/state.js
@@ -63,6 +63,8 @@ var stateSnapshots;
 
 let extensions;
 
+let utils;
+
 var listened = {};
 
 var p, hashArr = hash.replace("#", "").split("/");
@@ -144,6 +146,7 @@ module.exports = {
         advancedInfo = o.advancedInfo;
         meta = o.meta;
         layerTree = o.layerTree;
+        utils = o.utils;
         backboneEvents = o.backboneEvents;
         extensions = o.extensions;
         _self = this;
@@ -286,12 +289,16 @@ module.exports = {
                         } else {
                             // Set base layer to the first added one
                             setBaseLayer.init(baseLayer.getAvailableBaseLayers()[0].id);
-                            var extent = setting.getExtent();
-                            if (extent !== null) {
-                                cloud.get().zoomToExtent(extent);
-                            } else {
-                                cloud.get().zoomToExtent();
+
+                            if (!utils.parseZoomCenter(window.vidiConfig?.initZoomCenter)) {
+                                const extent = setting.getExtent();
+                                if (extent !== null) {
+                                    cloud.get().zoomToExtent(extent);
+                                } else {
+                                    cloud.get().zoomToExtent();
+                                }
                             }
+
                             if (window.vidiConfig.activeLayers.length > 0) {
                                 setLayers(false);
                             }
@@ -645,8 +652,8 @@ module.exports = {
      *
      * @returns {Promise}
      */
-    applyState: (state) => {
-
+    applyState: (state, ignoreInitZoomCenter = false) => {
+        console.log("ignoreInitZoomCenter", ignoreInitZoomCenter)
         if (LOG) console.log(`${MODULE_NAME}: applying state`, state);
 
         history.pushState(``, document.title, window.location.pathname + window.location.search);
@@ -694,7 +701,7 @@ module.exports = {
             };
 
             if ('map' in state) {
-                anchor.applyMapParameters(state.map).then(() => {
+                anchor.applyMapParameters(state.map, ignoreInitZoomCenter).then(() => {
                     applyStateToModules();
                 }).catch(error => {
                     console.error(error);
@@ -742,6 +749,12 @@ module.exports = {
     },
 
     setExtent: function () {
+        const arr = window.vidiConfig?.initZoomCenter ? utils.parseZoomCenter(window.vidiConfig.initZoomCenter) : null;
+        if (arr) {
+            hashArr[1] = arr.z;
+            hashArr[2] = arr.x;
+            hashArr[3] = arr.y;
+        }
         if (hashArr[1] && hashArr[2] && hashArr[3]) {
             p = geocloud.transformPoint(hashArr[2], hashArr[3], "EPSG:4326", "EPSG:3857");
             cloud.get().zoomToPoint(p.x, p.y, hashArr[1]);